extract root fields and computed fields components
