+++ console/src/components/Services/Data/Common/TooltipMessages.js
@@ -32,10 +32,6 @@ export const fkViolationOnDelete = `
 Action when the reference key is deleted
 `;
 
-export const changeViewRootFields = `
-Change the root fields for the view in the GraphQL API
-`;
-
 export const checkConstraintExpression = `
 Boolean expression that must be satisfied for all rows in the table. e.g. min_price >= 0 AND max_price >= min_price
 `;
