+++ console/src/components/Services/Data/TableModify/ModifyTable.js
@@ -30,6 +30,7 @@ import UniqueKeyEditor from './UniqueKeyEditor';
 import TriggerEditorList from './TriggerEditorList';
 import CheckConstraints from './CheckConstraints';
 import RootFields from './RootFields';
+import ComputedFields from './ComputedFields';
 import styles from './ModifyTable.scss';
 import { NotFoundError } from '../../../Error/PageNotFound';
 
@@ -38,13 +39,9 @@ import {
   getTableCheckConstraints,
   findTable,
   generateTableDef,
-  getTableCustomRootFields,
   getTableCustomColumnNames,
 } from '../../../Common/utils/pgUtils';
 import Tooltip from '../../../Common/Tooltip/Tooltip';
-import KnowMoreLink from '../../../Common/KnowMoreLink/KnowMoreLink';
-import ComputedFieldsEditor from './ComputedFieldsEditor';
-import ToolTip from '../../../Common/Tooltip/Tooltip';
 import {
   foreignKeyDescription,
   primaryKeyDescription,
@@ -71,8 +68,6 @@ class ModifyTable extends React.Component {
     const {
       tableName,
       allTables,
-      nonTrackableFunctions,
-      trackableFunctions,
       dispatch,
       migrationMode,
       readOnlyMode,
@@ -88,7 +83,6 @@ class ModifyTable extends React.Component {
       columnDefaultFunctions,
       schemaList,
       tableEnum,
-      rootFieldsEdit,
       postgresVersion,
     } = this.props;
 
@@ -158,56 +152,6 @@ class ModifyTable extends React.Component {
       );
     };
 
-    // if (table.primary_key.columns > 0) {}
-    const getTableRootFieldsSection = () => {
-      const existingRootFields = getTableCustomRootFields(table);
-
-      return (
-        <React.Fragment>
-          <h4 className={styles.subheading_text}>
-            Custom GraphQL Root Fields
-            <Tooltip
-              message={
-                'Change the root fields for the table in the GraphQL API'
-              }
-            />
-          </h4>
-          <RootFields
-            existingRootFields={existingRootFields}
-            rootFieldsEdit={rootFieldsEdit}
-            dispatch={dispatch}
-            tableName={tableName}
-          />
-          <hr />
-        </React.Fragment>
-      );
-    };
-
-    const getComputedFieldsSection = () => {
-      const allFunctions = nonTrackableFunctions.concat(trackableFunctions);
-
-      return (
-        <React.Fragment>
-          <h4 className={styles.subheading_text}>
-            Computed fields
-            <Tooltip
-              message={'Add a function as a virtual field in the GraphQL API'}
-            />
-            <KnowMoreLink href="https://hasura.io/docs/1.0/graphql/manual/schema/computed-fields.html" />
-          </h4>
-          <ComputedFieldsEditor
-            table={table}
-            currentSchema={currentSchema}
-            functions={allFunctions} // TODO: fix cross schema functions
-            schemaList={schemaList}
-            dispatch={dispatch}
-          />
-          <hr />
-        </React.Fragment>
-      );
-    };
-
-    // if (tableSchema.primary_key.columns > 0) {}
     return (
       <div className={`${styles.container} container-fluid`}>
         <TableHeader
@@ -253,10 +197,11 @@ class ModifyTable extends React.Component {
               postgresVersion={postgresVersion}
             />
             <hr />
-            {getComputedFieldsSection()}
+            <ComputedFields tableSchema={table} />
+            <hr />
             <h4 className={styles.subheading_text}>
               Primary Key &nbsp; &nbsp;
-              <ToolTip message={primaryKeyDescription} />
+              <Tooltip message={primaryKeyDescription} />
             </h4>
             <PrimaryKeyEditor
               tableSchema={table}
@@ -267,7 +212,7 @@ class ModifyTable extends React.Component {
             <hr />
             <h4 className={styles.subheading_text}>
               Foreign Keys &nbsp; &nbsp;
-              <ToolTip message={foreignKeyDescription} />
+              <Tooltip message={foreignKeyDescription} />
             </h4>
             <ForeignKeyEditor
               tableSchema={table}
@@ -280,7 +225,7 @@ class ModifyTable extends React.Component {
             <hr />
             <h4 className={styles.subheading_text}>
               Unique Keys &nbsp; &nbsp;
-              <ToolTip message={uniqueKeyDescription} />
+              <Tooltip message={uniqueKeyDescription} />
             </h4>
             <UniqueKeyEditor
               tableSchema={table}
@@ -296,7 +241,7 @@ class ModifyTable extends React.Component {
             <hr />
             <h4 className={styles.subheading_text}>
               Check Constraints &nbsp; &nbsp;
-              <ToolTip message={checkConstraintsDescription} />
+              <Tooltip message={checkConstraintsDescription} />
             </h4>
             <CheckConstraints
               constraints={getTableCheckConstraints(table)}
@@ -304,7 +249,8 @@ class ModifyTable extends React.Component {
               dispatch={dispatch}
             />
             <hr />
-            {getTableRootFieldsSection()}
+            <RootFields tableSchema={table} />
+            <hr />
             {getEnumsSection()}
             {untrackBtn}
             {deleteBtn}
@@ -340,8 +286,6 @@ ModifyTable.propTypes = {
 const mapStateToProps = (state, ownProps) => ({
   tableName: ownProps.params.table,
   allTables: state.tables.allSchemas,
-  nonTrackableFunctions: state.tables.nonTrackablePostgresFunctions || [],
-  trackableFunctions: state.tables.postgresFunctions || [],
   migrationMode: state.main.migrationMode,
   readOnlyMode: state.main.readOnlyMode,
   serverVersion: state.main.serverVersion,
