+++ console/src/components/Services/Data/TableModify/RootFields.js
@@ -1,106 +1,42 @@
 import React from 'react';
-import ExpandableEditor from '../../../Common/Layout/ExpandableEditor/Editor';
-import RootFieldEditor from '../Common/Components/RootFieldEditor';
-import { modifyRootFields, setCustomRootFields } from './ModifyActions';
-import { isEmpty } from '../../../Common/utils/jsUtils';
+import { connect } from 'react-redux';
+
+import RootFieldsEditor from './RootFieldsEditor';
+import Tooltip from '../../../Common/Tooltip/Tooltip';
 
 import styles from './ModifyTable.scss';
+import {
+  getTableCustomRootFields,
+  getTableName,
+} from '../../../Common/utils/pgUtils';
 
-const RootFields = ({
-  existingRootFields,
-  rootFieldsEdit,
-  dispatch,
-  tableName,
-}) => {
-  const setRootFieldsBulk = rf => {
-    dispatch(modifyRootFields(rf));
-  };
+const RootFields = props => {
+  const { tableSchema, rootFieldsEdit, dispatch } = props;
 
-  const onChange = (field, customField) => {
-    const newRootFields = {
-      ...rootFieldsEdit,
-      [field]: customField,
-    };
-    dispatch(modifyRootFields(newRootFields));
-  };
-
-  const collapsedLabel = () => {
-    const customRootFieldLabels = [];
-
-    Object.keys(existingRootFields).forEach(rootField => {
-      const customRootField = existingRootFields[rootField];
-      if (customRootField) {
-        customRootFieldLabels.push(
-          <React.Fragment>
-            {!isEmpty(customRootFieldLabels) && ', '}
-            <span className={styles.display_inline} key={rootField}>
-              <i>{rootField}</i> &rarr; {customRootField}
-            </span>
-          </React.Fragment>
-        );
-      }
-    });
-
-    if (isEmpty(customRootFieldLabels)) {
-      customRootFieldLabels.push('No root fields customised');
-    }
-
-    return <span>{customRootFieldLabels}</span>;
-  };
-
-  const editorExpanded = () => (
-    <RootFieldEditor
-      tableName={tableName}
-      rootFields={rootFieldsEdit}
-      selectOnChange={e => {
-        onChange('select', e.target.value);
-      }}
-      selectByPkOnChange={e => {
-        onChange('select_by_pk', e.target.value);
-      }}
-      selectAggOnChange={e => {
-        onChange('select_aggregate', e.target.value);
-      }}
-      insertOnChange={e => {
-        onChange('insert', e.target.value);
-      }}
-      insertOneOnChange={e => {
-        onChange('insert_one', e.target.value);
-      }}
-      updateOnChange={e => {
-        onChange('update', e.target.value);
-      }}
-      updateByPkOnChange={e => {
-        onChange('update_by_pk', e.target.value);
-      }}
-      deleteOnChange={e => {
-        onChange('delete', e.target.value);
-      }}
-      deleteByPkOnChange={e => {
-        onChange('delete_by_pk', e.target.value);
-      }}
-    />
-  );
-
-  const expandCallback = () => {
-    setRootFieldsBulk(existingRootFields);
-  };
-
-  const saveFunc = toggleEditor => {
-    dispatch(setCustomRootFields(toggleEditor));
-  };
+  const existingRootFields = getTableCustomRootFields(tableSchema);
 
   return (
-    <ExpandableEditor
-      editorExpanded={editorExpanded}
-      expandCallback={expandCallback}
-      collapsedLabel={collapsedLabel}
-      saveFunc={saveFunc}
-      property="custom-root-fields"
-      service="modify-table"
-      isCollapsable
-    />
+    <React.Fragment>
+      <h4 className={styles.subheading_text}>
+        Custom GraphQL Root Fields
+        <Tooltip message="Change the root fields for the table/view in the GraphQL API" />
+      </h4>
+      <RootFieldsEditor
+        existingRootFields={existingRootFields}
+        rootFieldsEdit={rootFieldsEdit}
+        dispatch={dispatch}
+        tableName={getTableName(tableSchema)}
+      />
+      <hr />
+    </React.Fragment>
   );
 };
 
-export default RootFields;
+const mapStateToProps = (state, ownProps) => {
+  return {
+    tableSchema: ownProps.tableSchema,
+    rootFieldsEdit: state.tables.modify.rootFieldsEdit,
+  };
+};
+
+export default connect(mapStateToProps)(RootFields);
